requirements.txt
requests
azure-eventhub
azure-servicebus